Displays the Ajax option even if WP_CACHE is false
As discussed on the support forum, there are caching methods that doesn't require WP_CACHE to be true. Think of Nginx FastCGI cache, or to Redis. To understand the importance for your plugin to be compatible with Nginx just think at the wordpress.org server, that is right Nginx. This will enable the Ajax options for everyone, even if WP_CACHE is false, that is not so bad.

Thanks!